Allow connection to generic lambdas and overloaded function objects
Implemented by being more pragmatic, it simplifies things a lot and should also
have an observable effect on compilation times.

The earlier, analytic approach to callable types decomposition was problematic
in the sense that not every callable is a function of function object. Generic
lambdas and overloaded function object call operators spring to mind most notably.

The new approach consist in checking whether the supplied callable can be called
with the signal argument types, which is much easier and foolproof.