Improve signal emission performance by avoid unnecessary slot copies
We use a copy on write mechanism that will limit copies of slots to when an
access pattern makes it absolutely unavoidable.
This should provide substantial speed ups in real code bases, because signal
mutation (creation, connection) tends to happen once at the beginning of a
program, whereas emission happens when everything has been set up.