Implement safer disconnection of slots
In multithreaded contexts, sigslot does not invoke slots under a lock
in order to avoid deadlocks under certains scenarii. This is also a
net gain for concurrency. This makes it however far more difficult to
reason about the lifetime of callables connected to a signal.

One such case is when slots disconnection happens concurrently with signal
emission and slots invocation. Disconnection cannot be immediate because
the underlying callables may be in use. For callables relying on an object
instance, knowing when the instance could be safely destructed proved to
be difficult.

To mitigate this issue, the library offers the concept of trackable objects,
such as std::shared_ptr, to manage the lifetime of objects. Those are used
to extend the lifetime of objects while a slot is being invoked.
Unfortunately this put the burden of lifetime management entirely on the
user side and proved difficult to use in practice.

This is proof of concep introduces internal tracking of slots invocations,
which makes it possible to defer slots destruction until they are not used
anymore.

For multithreaded applications, this change makes it far easier to manage
the lifetime of callables:

- calls to `signal_base::disconnect` and `signal_base_disconnect_all`
- calls to connection::disconnect(true)
- `scoped_connection` and `observer` destruction

now all give the guarantee that the disconnected slots are not in use and
will not be used anymore.