Add exception handling to signal emission
The `operator()` method inside `signal.hpp` and new test cases have been added to handle the functionality of exceptions when signals are emitted. Now the function returns a vector of `std::exception_ptr`s for any exceptions that occurred during the signal emission.

Since this is an ABI breaking change, the version has been updated to 2.0.0.